+++ Src/ImplicitNullability.Plugin.Tests/test/data/Integrative/AnnotationCopyingTests/CheckParamNullAction.cs
@@ -1,0 +1,9 @@
+﻿namespace ImplicitNullability.Plugin.Tests.test.data.Integrative.AnnotationCopyingTests
+{
+    public class Sample
+    {
+        public void Method(string{caret} a)
+        {
+        }
+    }
+}