+++ Src/ImplicitNullability.Plugin/ReSharperCustomImplicitNullabilityProvider.cs
@@ -3,23 +3,27 @@ using JetBrains.Annotations;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CodeAnnotations;
 using JetBrains.Util;
+#if RESHARPER20171
+using ICustomImplicitNullabilityProvider = JetBrains.ReSharper.Psi.CodeAnnotations.ICustomCodeAnnotationProvider;
+
+#endif
 
 namespace ImplicitNullability.Plugin
 {
     /// <summary>
-    /// An implementation of ReSharper's <see cref="ICustomCodeAnnotationProvider"/> extension point which uses
+    /// An implementation of ReSharper's <see cref="ICustomImplicitNullabilityProvider"/> extension point which uses
     /// the <see cref="ImplicitNullabilityProvider"/> to implement the Implicit Nullability rules.
     /// 
     /// Note that these "custom" providers are called by R# at last (basically if there are no (inherited) nullability attributes).
     /// </summary>
     [PsiComponent]
-    public class ImplicitNullabilityCustomCodeAnnotationProvider : ICustomCodeAnnotationProvider
+    public class ReSharperCustomImplicitNullabilityProvider : ICustomImplicitNullabilityProvider
     {
-        private static readonly ILogger Logger = JetBrains.Util.Logging.Logger.GetLogger(typeof(ImplicitNullabilityCustomCodeAnnotationProvider));
+        private static readonly ILogger Logger = JetBrains.Util.Logging.Logger.GetLogger(typeof(ReSharperCustomImplicitNullabilityProvider));
 
         private readonly ImplicitNullabilityProvider _implicitNullabilityProvider;
 
-        public ImplicitNullabilityCustomCodeAnnotationProvider(ImplicitNullabilityProvider implicitNullabilityProvider)
+        public ReSharperCustomImplicitNullabilityProvider(ImplicitNullabilityProvider implicitNullabilityProvider)
         {
             Logger.Verbose(".ctor");
             _implicitNullabilityProvider = implicitNullabilityProvider;